+++ 2021/day01/lib.h
@@ -6,7 +6,7 @@
 
 namespace day01
 {
-    size_t count_increases(const std::vector<std::string> &input, size_t window_size);
+    uint32_t count_increases(const std::vector<std::string> &input, size_t window_size);
 }
 
 #endif
