+++ 2021/day02/lib.cc
@@ -1,6 +1,6 @@
 #include "lib.h"
 
-size_t day02::run_part1(const std::vector<std::string> &input)
+int32_t day02::run_part1(const std::vector<std::string> &input)
 {
     int32_t horizontal = 0;
     int32_t vertical = 0;
@@ -9,17 +9,17 @@ size_t day02::run_part1(const std::vector<std::string> &input)
     {
         if (line.find("forward ") == 0)
         {
-            int32_t value = std::stol(line.substr(8));
+            int32_t value = std::stoi(line.substr(8));
             horizontal += value;
         }
         else if (line.find("down ") == 0)
         {
-            int32_t value = std::stol(line.substr(5));
+            int32_t value = std::stoi(line.substr(5));
             vertical += value;
         }
         else if (line.find("up ") == 0)
         {
-            int32_t value = std::stol(line.substr(3));
+            int32_t value = std::stoi(line.substr(3));
             vertical -= value;
         }
     }
@@ -27,7 +27,7 @@ size_t day02::run_part1(const std::vector<std::string> &input)
     return horizontal * vertical;
 }
 
-size_t day02::run_part2(const std::vector<std::string> &input)
+int32_t day02::run_part2(const std::vector<std::string> &input)
 {
     int32_t horizontal = 0;
     int32_t vertical = 0;
@@ -37,18 +37,18 @@ size_t day02::run_part2(const std::vector<std::string> &input)
     {
         if (line.find("forward ") == 0)
         {
-            int32_t value = std::stol(line.substr(8));
+            int32_t value = std::stoi(line.substr(8));
             horizontal += value;
             vertical += value * aim;
         }
         else if (line.find("down ") == 0)
         {
-            int32_t value = std::stol(line.substr(5));
+            int32_t value = std::stoi(line.substr(5));
             aim += value;
         }
         else if (line.find("up ") == 0)
         {
-            int32_t value = std::stol(line.substr(3));
+            int32_t value = std::stoi(line.substr(3));
             aim -= value;
         }
     }
